Allow force_ssl override in production via ENV
While SSL is generally a good idea, some non-Heroku use cases for Heaven
may prefer disabling SSL. For example, it may be run behind a
firewall with a proxy server terminating the SSL connection. In this
case, it may be acceptable to run Heaven in a trusted environment
without having to configure SSL certificates.

The force_ssl setting is still enabled by default in the production and
staging environments, as it should be for any public-facing Heaven
installation (e.g., Heroku). The default behavior can be overridden
by setting the `FORCE_SSL` environment variable to 'false'